use redteam retry logic
